workspace/skills/secure-gmail/requirements.txt
python-dotenv>=1.0.0
composio-core>=0.4.0
